Dj_Configuration/requirements.txt
asgiref==3.8.1
black==24.4.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
pycparser==2.22
PyJWT==2.8.0
requests==2.32.2
requests-oauthlib==2.0.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.0
urllib3==2.2.1
